Rename input_store files
