+++ Anvil/anvil.dm.cpp
@@ -61,7 +61,7 @@ Anvil::Anvil(spkt::window* window)
     d_window->set_cursor_visibility(true);
 
     d_scene = std::make_shared<spkt::scene>(); 
-    spkt::add_singleton(d_scene->registry);   
+    spkt::add_singleton(d_scene->registry, d_window);   
     spkt::load_registry_from_file(d_sceneFile, d_scene->registry);
     d_activeScene = d_scene;
 }
@@ -200,7 +200,7 @@ void Anvil::on_render()
             if (ImGui::MenuItem("Run")) {
                 d_activeScene = std::make_shared<spkt::scene>();
 
-                spkt::add_singleton(d_activeScene->registry);
+                spkt::add_singleton(d_activeScene->registry, d_window);
                 spkt::copy_registry(d_scene->registry, d_activeScene->registry);
 
                 d_activeScene->systems = {
