+++ Sprocket/Scene/scene.cpp
@@ -4,7 +4,7 @@
 #include <Sprocket/Core/events.h>
 #include <Sprocket/Scene/ecs.h>
 #include <Sprocket/Scene/Systems/input_system.h>
-#include <Sprocket/Utility/InputProxy.h>
+#include <Sprocket/Utility/input_store.h>
 
 #include <memory>
 #include <ranges>
